+++ server/src/main/java/de/njsm/stocks/server/v2/web/LocationEndpoint.java
@@ -121,7 +121,7 @@ public class LocationEndpoint extends Endpoint {
                                    @QueryParam("id") int id,
                                    @QueryParam("version") int version,
                                    @FormParam("description") String description) {
-        if (isValid(id, "id") && isValidVersion(version, "version") && isValid(description, "description")) {
+        if (isValid(id, "id") && isValidVersion(version, "version") && isValidOrEmpty(description, "description")) {
             manager.setPrincipals(getPrincipals(request));
             StatusCode result = manager.setDescription(new LocationForSetDescription(id, version, description));
             return new Response(result);